Reorganize student assessment instance page code
